requirements.txt
PyJWT~=2.9.0
PyMySQL~=1.1.1
bcrypt~=4.2.0
Flask~=3.0.3
Flask-JWT-Extended~=4.6.0
Werkzeug~=3.0.4
route~=0.1.4
Routes~=2.5.1
python-dotenv~=1.0.1
Flask-Migrate~=4.0.7